+++ dadb/src/test/kotlin/dadb/DadbTest.kt
@@ -24,18 +24,24 @@ import okio.source
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
 import java.io.ByteArrayInputStream
+import java.io.FileInputStream
 import java.net.Socket
 import java.nio.charset.StandardCharsets
-import java.util.Random
+import java.util.*
 import java.util.concurrent.Callable
 import java.util.concurrent.Executors
 import java.util.concurrent.Future
 import java.util.concurrent.TimeUnit
+import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Ignore
 import kotlin.test.Test
 
-internal class DadbTest : BaseConcurrencyTest() {
+internal abstract class DadbTest : BaseConcurrencyTest() {
+
+    private val remotePath = "/data/local/tmp/hello"
+
+    private val remotePathWithCJK = "/data/local/tmp/你好こんにちは"
 
     @JvmField
     @Rule
@@ -46,7 +52,14 @@ internal class DadbTest : BaseConcurrencyTest() {
     @BeforeTest
     fun setUp() {
         temporaryFolder.create()
-//        killServer()
+    }
+
+    @AfterTest
+    internal fun tearDown() {
+        localEmulator { dadb ->
+            dadb.shell("rm -f $remotePath")
+            dadb.shell("rm -f $remotePathWithCJK")
+        }
     }
 
     @Test
@@ -111,7 +124,6 @@ internal class DadbTest : BaseConcurrencyTest() {
     fun adbPush_basic() {
         localEmulator { dadb ->
             val content = randomString()
-            val remotePath = "/data/local/tmp/hello"
 
             val source = ByteArrayInputStream(content.toByteArray()).source()
             dadb.push(source, remotePath, 439, System.currentTimeMillis())
@@ -147,7 +159,6 @@ internal class DadbTest : BaseConcurrencyTest() {
     @Test
     fun adbPull_largeFile() {
         localEmulator { dadb ->
-            val remotePath = "/data/local/tmp/hello"
             val sizeMb = 100
 
             dadb.shell("fallocate -l ${sizeMb}M $remotePath")
@@ -161,10 +172,24 @@ internal class DadbTest : BaseConcurrencyTest() {
     }
 
     @Test
+    fun adbPush_pathWithCJK() {
+        localEmulator { dadb ->
+            val content = randomString()
+            val localSrcFile = temporaryFolder.newFile().apply { writeText(content) }
+
+            dadb.push(localSrcFile, remotePathWithCJK, 439, System.currentTimeMillis())
+
+            val localDstFile = temporaryFolder.newFile()
+            dadb.pull(localDstFile, remotePathWithCJK)
+
+            assertThat(localDstFile.readText()).isEqualTo(content)
+        }
+    }
+
+    @Test
     fun adbPush_file() {
         localEmulator { dadb ->
             val content = randomString()
-            val remotePath = "/data/local/tmp/hello"
             val localSrcFile = temporaryFolder.newFile().apply { writeText(content) }
 
             dadb.push(localSrcFile, remotePath, 439, System.currentTimeMillis())
@@ -186,11 +211,21 @@ internal class DadbTest : BaseConcurrencyTest() {
     }
 
     @Test
+    fun installStream() {
+        localEmulator { dadb ->
+            val inputStream = FileInputStream(TestApk.FILE).source()
+            dadb.install(inputStream, TestApk.FILE.length())
+            val response = dadb.shell("pm list packages ${TestApk.PACKAGE_NAME}")
+            assertShellResponse(response, 0, "package:${TestApk.PACKAGE_NAME}\n")
+        }
+    }
+
+    @Test
     fun installMultiple() {
         localEmulator { dadb ->
             dadb.installMultiple(TestApk.SPLIT_FILES)
-            val response = dadb.shell("pm list packages ${TestApk.SPLIT_PACKAGE_NAME}")
-            assertShellResponse(response, 0, "package:${TestApk.SPLIT_PACKAGE_NAME}\n")
+            val response = dadb.shell("pm path ${TestApk.SPLIT_PACKAGE_NAME} | wc -l")
+            assertShellResponse(response, 0, "${TestApk.SPLIT_FILES.size}\n")
         }
     }
 
@@ -284,13 +319,7 @@ internal class DadbTest : BaseConcurrencyTest() {
         }
     }
 
-    private fun localEmulator(body: (dadb: Dadb) -> Unit) {
-        val socket = Socket("localhost", 5555)
-        val keyPair = AdbKeyPair.readDefault()
-        val connection = AdbConnection.connect(socket, keyPair)
-        TestDadb(connection).use(body)
-        connection.ensureEmpty()
-    }
+    protected abstract fun localEmulator(body: (dadb: Dadb) -> Unit)
 
     private fun readSocketAsync(host: String, port: Int): Future<String> {
         return executor
@@ -312,7 +341,7 @@ internal class DadbTest : BaseConcurrencyTest() {
         val future = executor.submit {
             dadb.shell("echo -e '$message' | nc -lp $port")
         }
-        Thread.sleep(100)
+        Thread.sleep(500)
 
         return future
     }
@@ -320,15 +349,4 @@ internal class DadbTest : BaseConcurrencyTest() {
     private fun randomString(): String {
         return "${Random().nextDouble()}"
     }
-}
-
-private class TestDadb(
-    private val connection: AdbConnection,
-) : Dadb {
-
-    override fun open(destination: String) = connection.open(destination)
-
-    override fun supportsFeature(feature: String) = connection.supportsFeature(feature)
-
-    override fun close() = connection.close()
 }